4단계 - 로또(수동) (#2682)
* refactor: 로또 가격을 Ticket 이 가지게 변경

* refactor: 1~45번 생성은 한번만 이뤄지게 변경

* style: 로또볼 생성사 파라미터명 수정

* feat: 보너스 번호와 2등 조건 추가 및 그에 따른 결과 확인 로직 수정

* style: 사용되지 않는 코드 제거

* refactor: static 메소드만 가지는 클래스 private 생성자 추가 및 인풋 구문자 매직스트링으로 변경

* refactor: 보너스 룰 확인 조건문 private 메소드로 포장

* refactor: 로또볼 번호 비교 시 equals 을 사용하게 변경

* test: 당첨금을 가지는 당첨조건 조회 메소드 테스트 추가

* test: 당첨 티켓 수 구하기 테스트 당첨조건별로 독립적으로 테스트하게 변경.

* test: WinningCondition 테스트 코드 정리

* feat: 로또마킹지와 수동 로또기계 구현

* feat: 티켓 박스 합치기 기능 추가

* feat: 투입금액 빼기 기능 추가

* feat: input view 에 수동 로또 추가

* feat: result view 에 수동 로또 장수도 추가

* feat: 수동 로또도 발급받게 추가

* docs: 로또 발급 방식 설명 추가

* refactor: prizeMoney long 타입으로 변경

* refactor: LottoBall 미리 생성하여 재사용하게 변경

* chore: lottomachine 패키지 정리

* refactor: ManualLottoMachine 에서 미리 마킹된 MarkingPaper 를 생성자 주입으로 받게 변경

* feat: 게임 실행 컨트롤러 LottoController 추가

+++ src/main/java/lotto/domain/winner/WinningCondition.java
@@ -18,19 +18,19 @@ public enum WinningCondition {
     ;
 
     private final int matchCount;
-    private final int prizeMoney;
+    private final long prizeMoney;
 
     private final boolean requiredBonus;
     private final boolean bonusMatch;
 
-    WinningCondition(int matchCount, int prizeMoney) {
+    WinningCondition(int matchCount, long prizeMoney) {
         this.matchCount = matchCount;
         this.prizeMoney = prizeMoney;
         this.requiredBonus = false;
         this.bonusMatch = false;
     }
 
-    WinningCondition(int matchCount, int prizeMoney, boolean requiredBonus, boolean bonusMatch) {
+    WinningCondition(int matchCount, long prizeMoney, boolean requiredBonus, boolean bonusMatch) {
         this.matchCount = matchCount;
         this.prizeMoney = prizeMoney;
         this.requiredBonus = requiredBonus;
@@ -60,7 +60,7 @@ public enum WinningCondition {
         return matchCount;
     }
 
-    public int getPrizeMoney() {
+    public long getPrizeMoney() {
         return prizeMoney;
     }
